fix: shift renderBrush functionality to a mixin
