+++ src/mixins/index.js
@@ -4,6 +4,7 @@ import circleSegmentationMixin from './segmentation/circleSegmentationMixin.js';
 import polylineSegmentationMixin from './segmentation/polylineSegmentationMixin.js';
 import freehandSegmentationMixin from './segmentation/freehandSegmentationMixin.js';
 import rectangleSegmentationMixin from './segmentation/rectangleSegmentationMixin.js';
+import renderBrushMixin from './segmentation/renderBrushMixin.js';
 
 export default {
   activeOrDisabledBinaryTool,
@@ -12,4 +13,5 @@ export default {
   polylineSegmentationMixin,
   freehandSegmentationMixin,
   rectangleSegmentationMixin,
+  renderBrushMixin,
 };
